Adding code to controller concern
