+++ app/models/appeal.rb
@@ -9,7 +9,6 @@ require "securerandom"
 
 # rubocop:disable Metrics/ClassLength
 class Appeal < DecisionReview
-  # include AppealConcern
   include BeaamAppealConcern
   include BgsService
   include Taskable
